python/requirements.txt
pywin32==306
docx2pdf==0.1.8
python-docx==1.1.0
docxtpl==0.16.7
Pillow==10.1.0
